rename MpiApi to MpiServer
